Fix C src warnings raised as compiled errors

+++ pandas/_libs/src/librdata/rdata_write.c
@@ -167,7 +167,7 @@ static rdata_error_t rdata_write_string(
     if (retval != RDATA_OK)
         goto cleanup;
 
-    ssize_t len = string ? strlen(string) : -1;
+    ssize_t len = string ? (ssize_t)strlen(string) : -1;
 
     retval = rdata_write_integer(writer, len);
     if (retval != RDATA_OK)
